+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 # 0.15.2 - 19 Jul, 2018
 
+- Introduced support for specifying multiple directories for the Migrator
+
 ### Changes:
 
 - Rolled back changes introduced by #2542, in favor of opt-in behavior by adding a precision option in `date` / `timestamp` / `datetime` / `knex.fn.now` (#2715, #2721)